+++ stories/templates/checkbox/default.template.js
@@ -1,3 +1,8 @@
 export const template = `
-  <st-checkbox v-model="isChecked" :disabled="disabled" :readonly="readonly">{{ innerText }}</st-checkbox>
+  <st-checkbox v-model="isChecked"
+               :disabled="disabled"
+               :readonly="readonly"
+               :indeterminate="indeterminate">
+    {{ innerText }}
+  </st-checkbox>
 `;